refactor: lottoNumbersFactory lottoTicketFactory 로 변환
- 로또 번호들을 생성해주지 않고 바로 ticket 을 발급하게 변경
