+++ src/main/java/lotto/domain/LottoTicketFactory.java
@@ -8,19 +8,19 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class LottoNumbersFactory {
+public class LottoTicketFactory {
     private static final String COMMA = ", ";
 
-    private LottoNumbersFactory() {
+    private LottoTicketFactory() {
     }
 
-    public static List<Integer> createLottoNumbers() {
+    public static LottoTicket createLottoTicket() {
         List<Integer> allLottoNumbers = createAllLottoNumbers();
         Collections.shuffle(allLottoNumbers);
         List<Integer> lottoNumbers = allLottoNumbers.subList(
             LottoNumbersSize.LOTTO_NUMBERS_ZERO_SIZE.getSize(), LottoNumbersSize.LOTTO_NUMBERS_SIZE.getSize());
         Collections.sort(lottoNumbers);
-        return lottoNumbers;
+        return new LottoTicket(lottoNumbers);
     }
 
     private static List<Integer> createAllLottoNumbers() {
@@ -30,14 +30,14 @@ public class LottoNumbersFactory {
             .collect(Collectors.toList());
     }
 
-    public static List<Integer> createManualLottoNumbers(String inputNumbers) {
+    public static LottoTicket createManualLottoTicket(String inputNumbers) {
         List<Integer> numbers = new ArrayList<>();
         String[] inputNumbersBySplit = inputNumbers.split(COMMA);
         for (String number : inputNumbersBySplit) {
             validatePositiveNumber(number);
             numbers.add(Integer.parseInt(number));
         }
-        return numbers;
+        return new LottoTicket(numbers);
     }
 
     private static void validatePositiveNumber(String number) {
